backend/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
dj-database-url==2.3.0
psycopg2-binary==2.9.10
celery==5.5.2
redis==5.2.1
neo4j==5.28.1
django-filter==25.2